🔧 Fix 5 PHPStan type errors and improve testing infrastructure
Type Safety Fixes (5 errors resolved):
1. Remove unused $admin_hide_security property (never read)
2. Add null check before rtrim() in remove_version_query_string()
3. Update check_login_lockout() return type to include null
4. Cast get_remaining_lockout_time() return to int
5. Handle preg_replace() null return properly

PHPStan Configuration Improvements:
- Create tests/phpstan-constants.php for plugin constants
- Remove deprecated checkMissingIterableValueType option
- Remove deprecated checkGenericClassInNonGenericObjectType option
- Clean up unused ignore patterns (4 patterns removed)
- Use error identifiers instead of deprecated config options
- Add bootstrap file for constants loading

Testing Infrastructure:
- Add scripts/run-quality-checks.sh for local CI/CD matching
- Comprehensive quality checks matching GitHub Actions
- Two-stage testing: PHPStan (standalone) + PHPUnit (WordPress)
- Document testing strategy in README.md
- Security plugin requires real WordPress environment testing

Why Two-Stage Testing:
- PHPStan: Fast type checking WITHOUT WordPress (30s)
- PHPUnit: Real security validation WITH WordPress (2-3m)
- Both required: Type safety + actual security feature validation

All checks now pass:
✅ PHPStan Level 8 (no errors)
✅ WordPress Test Suite ready
✅ Local testing matches CI/CD exactly

+++ src/Security/LoginSecurity.php
@@ -282,7 +282,7 @@ class LoginSecurity {
 	 * @param \WP_User|\WP_Error|null $user User object or error
 	 * @param string                  $username Username
 	 * @param string                  $password Password
-	 * @return \WP_User|\WP_Error
+	 * @return \WP_User|\WP_Error|null
 	 */
 	public function check_login_lockout( $user, string $username, string $password ) {
 		// Skip if no username/password provided
@@ -583,7 +583,7 @@ class LoginSecurity {
 		);
 
 		if ( $transient_timeout ) {
-			return max( 0, $transient_timeout - time() );
+			return (int) max( 0, $transient_timeout - time() );
 		}
 
 		return 0;
